pyproject.toml
[tool.poetry]
name = "adraft"
version = "0.1.0-alpha"
description = "This is only a draft."

license = "MIT"

authors = [
    "I.G. Chernyakov <i.g.chernyakov@yandex.ru>"
]

readme = 'README.rst'

repository = "https://github.com/i-g-chernyakov/adraft"
homepage = "https://github.com/i-g-chernyakov/adraft"

keywords = [
    "django",
    "adraft"
]

classifiers = [
    "Development Status :: 1 - Planning",
    "Environment :: Web Environment",
    "Framework :: Django :: 2.1",
    "Intended Audience :: Customer Service",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: Implementation :: CPython",
]

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.1"
psycopg2-binary = { version = "^2.7.5", optional = true }
mysqlclient = { version = "^1.3", optional = true }
dj-database-url = "^0.5.0"
goodconf = "~1.0.0"
python-json-logger = "^0.1.9"
whitenoise = "^4.0"
"ruamel.yaml" = "^0.15.61"
python-magic = "^0.4.15"
Markups = "^3.0.0"
djangorestframework = "^3.8.2"
cffi = "^1.11"
django-sendfile = "^0.3.11"


[tool.poetry.extras]
mysql = ["mysqlclient"]
pgsql = ["psycopg2-binary"]


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^1.9"
pytest-cov = "^2.5"
pytest-isort = "^0.2.0"
pytest-flake8 = "^1.0"
pytest-randomly = "^1.2"
pytest = "^3.6"
flake8-builtins = "^1.4"
flake8-commas = "^2.0"
flake8-quotes = "^1.0"
flake8-comprehensions = "^1.4"
flake8-blind-except = "^0.1.1"
flake8-docstrings = "^1.3"
flake8-string-format = "^0.2.3"
flake8-coding = "^1.3"
flake8-module-name = "^0.1.5"
flake8-bugbear = "^18.2"
flake8-pytest = "^1.3"
pep8-naming = "^0.7.0"
mypy = "^0.610.0"
sphinx = "^1.7"
sphinx-autodoc-typehints = "^1.3"
sphinxcontrib-napoleon = "^0.6.1"
doc8 = "^0.8.0"
m2r = "^0.1.15"
sphinx_readable_theme = "^1.3"
tox = "^3.0"


[tool.poetry.scripts]
manage = 'adraft:config.django_manage'
"manage.py" = 'adraft:config.django_manage'
generate-config = 'adraft:generate_config'
